requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
attrs==19.1.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
configparser==3.5.0
cycler==0.10.0
decorator==4.4.0
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
hyperopt==0.1.2
ipython==5.5.0
ipython-genutils==0.2.0
joblib==0.11
JsonForm==0.0.2
jsonschema==3.0.1
JsonSir==0.0.2
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==2.1.1
mock==3.0.5
networkx==2.2
numpy==1.16.5
nvidia-ml-py==7.352.0
pandas==0.22.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.8.0
pymongo==3.9.0
pyparsing==2.2.0
pyrsistent==0.15.2
python-dateutil==2.6.1
Python-EasyConfig==0.1.7
pytools==2017.6
pytz==2018.3
PyWavelets==0.5.1
PyYAML==5.1
Resource==0.2.1
scandir==1.7
scikit-cuda==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.2.2
simplegeneric==0.8.1
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.7
termcolor==1.1.0
torch==1.0.1
torchvision==0.2.2
tqdm==4.36.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
